.github/workflows: restore 2 hours GitHub Action timeout when testing
The GitHub Action not terminating issue (#1531) seems to occur
regardless of the timeout. This is discussed in issue #1531.

For this reason, this commit restores a timeout, setting it to 2 hours,
thus aligning the GitHub Action timeout with the timeout imposed on the
'go test' call when commit.go is executed.

For #1531.