assets: replace Sentry feedback button with a link to `CONTRIBUTING.md`
This commit removes Sentry's feedback button, as its behavior appears to
be unreliable, with issues related to previewing and sending screenshots
occurring in some browsers.

In place of Sentry's system, this commit adds a link to the
`CONTRIBUTING.md` file in our repository, so that the user is guided to
report any bugs encountered through the issues system instead.