core/connectors/httpclient: document the case when the state is nil
This commit documents the use case of the 'httpclient.HTTP' type without
providing a state, which is useful for tests that don't require OAuth
and can therefore instantiate this type more easily, legitimizing it as
a valid case and ensuring its support.

Additionally, a parameter consistency check has been added.