commit, .github/workflows: remove option '-x' and simplify
This commit removes the '-x' (explicit) option from 'commit/commit.go'
making it the default behavior, as its presence complicated the script
overall, even though it was almost always used in practice, making it an
unnecessary general complication.

This also simplifies the execution of Go tests, as it's no longer
necessary to support module-level (rather than package-level) testing.

Note that even though '-x' has been removed and is now the default
behavior, it's still possible to run the script silently, if needed, by
redirecting stdout/stderr in the shell.