cmd: fix session cookie `Secure` flag when TLS is terminated by proxy
If Meergo is running behind a proxy that handles TLS termination, the 
session cookie is returned with the Secure flag set to false.

This happens because the Secure attribute is currently determined solely
by the MEERGO_HTTP_TLS_ENABLED environment variable. However, it should
be  based on the URL defined in MEERGO_HTTP_EXTERNAL_URL.

This change makes it use both values for a more accurate determination.