assets: filter selectable actions based on target in metrics page
This commit ensures that on the metrics page, the list of actions
selectable to display metrics in a granular manner is filtered based on
the selected target, so that for example if the selected target is
"Event", only actions on events are shown in the list.

Also, when you change targets, the currently selected action, if any, is
remembered so that if you then return to the previous target, that
action is automatically reselected in the filter.