core/events/collector/sender: stop retrying events on unknown errors
Previously, the sender would retry sending events whenever the
SendEvents method returned an error that wasn't a
`meergo.EventsError`.

However, those events would remain in the queue and be retried rapidly
in tight loops, without any delay.

This change removes those events from the queue immediately and logs
the error to the metrics system for each individual event.

A future improvement will allow the HTTP client to signal whether the
events should be retried and specify the appropriate delay.