assets: list the icons used by the admin via file
This commit adds a file listing all icons used by the admin.

This was necessary because in some cases icons are passed dynamically to
components, even via complex data structures, and the script that was
responsible for vendoring therefore was unable to locate them, leading
then to missing icons in the production build.

From this change onward, following changes to the UI, it will then be
necessary to consider whether to manually add or remove icons from the
file that lists them.

Resolves #1588.