connectors/mixpanel: correct max-body-size check in `sendEvents` method
The `sendEvents` method previously allowed sending a request body that
could exceed Mixpanel's maximum allowed size in bytes.

This change corrects that behavior and improves the check on the
number of events per request by preventing unnecessary calls to the
`Skip` method once the maximum number of events has been reached.

The logic for these two checks is also now consistent with that used
in other connectors.

Additionally, it fixes the import order in the test files.