all: send stats to a Meergo instance
This commit causes Meergo to send state change information to a remote
Meergo instance.

Currently, the Meergo SDK has been embedded in this repository; in the
future it will be removed and the SDK that is on GitHub will be used
directly.

All this happens without sending personal and/or sensitive data, and
always in compliance with the user's telemetry data sending settings,
who can choose to send nothing.

This commit is a first step; future commits may refine this part to make
it more useful/usabile.

For more specific information on the data sent and how to configure it,
see the changes to the DEVELOPERS.md file.