all: switch `/connections/:id/identities` from POST to GET
This commit changes the HTTP method for the
`/connections/:id/identities` endpoint from POST to GET to better
reflect its read-only nature. This means that parameters, currently
passed in the body as JSON, are encoded in the query string.

No documentation updates were needed, as it already reflected the GET
method and query string usage.