cmd: add a minimal CLI to Meergo
Currently, Meergo completely lacks a CLI. This means that running the
binary with the --help option does not return an error and Meergo
starts, even if you just want to see the CLI documentation.

This commit adds a minimal CLI that only prints the help. This allows
handling the standard case "meergo --help". It's definitely better than
nothing.

Later we will evaluate if and which flags to add.

For #1513.