all: add the `context.os.other` event field, like `browser.other`
This commit adds the 'context.os.other' field for consistency, with
behavior similar to 'context.browser.other'.

Note that this part will probably be developed later in future commits,
to change the behavior of "other" (both for browsers and OS) to make it
clearer and more usable.

It also includes the SQL script needed to update the PostgreSQL
warehouse before using it with this new version of Meergo.