meergo,core/connectors/httpclient: use `httpclient.Client` for OAuth
This change updates the `retrieveOAuthToken` method to use an
`httpclient.Client` for requests. This way, the connector’s rate
limiting and retry policies are properly applied.

Also, if the connection supports OAuth and the request has a retriable
body, the request will now be automatically retried on a 401 response.

To support this, a new `Unauthorized` value has been added to the
`meergo.FailureReason` type.