cmd, core: make the MaxMind db path configurable through env var
This commit ensures that the MaxMind database path is no longer
hard-coded in the code, but is now configurable via an environment
variable. If not set, the geolocation information are not automatically
added to the event.

Regarding this, compared to the current behavior where no error is
returned if the file doesn't exist, this commit ensures that an error is
returned if the path explicitly specified via env var doesn't exist.

This makes the whole usage more clear and explicit.