core/events/collector: improve decoding of browser and os name
This commit improves the decoding behavior of context.browser and
context.os.

Specifically, Meergo attempts to normalize the name, and if it fails, it
places it in the context.browser.other/context.os.other field. For more
details about this, see the documentation for these fields, both for
ingestion and reading, which has been changed by this commit.

Furthermore, it ensures that the user agent is parsed only if necessary;
otherwise, if both the browser and the OS are already present, it is not
parsed to determine them, as it happens on 'main' instead.

It also adds tests for the new behavior.