all: always include the 'user' field in events and document it
Currently, on main, the two event schemas—the one "incoming" to Meergo,
without a user, and the one from the warehouse, possibly with an
associated user—are merged into a single schema, which causes logical
validation issues and confusion.

However, instead of splitting the schemas again, the decision was made
to keep a single schema that always includes the 'user' field, although
in some cases it is never returned.

So, this commit implements this behavior and document it.