all: extend and standardize the previous ID field in events
For consistency, we decided to include the previous ID (mandatory for
alias-type events) in the events schema, save it in the warehouse,
document it, and return it, even though it is not currently used by the
Meergo model to resolve identities.

This is solely for the purpose of simplification and compatibility.

Resolves #1599.