connectors/klaviyo: pre-validate events and improve error handling
When sending events to Klaviyo, if one or more events fail Klaviyo's 
validation, none of the events in the request are accepted and the 
response body does not specify which events were invalid, making it 
impossible to report the specific error.

This change validates events before sending them to Klaviyo and 
discards any that are invalid.

Additionally, it extends the event type schema with new properties and
handles cases where the response is not successful.

For #1604