cmd, doc: implement a first version of an MCP server
This first version of the MCP server allows interaction with an LLM
application via Meergo to retrieve information about the user and event
schemas.

Clearly, this is just an initial commit that lays the foundation for an
infrastructure to be expanded over time.

For more details and on how to use the MCP server, refer to the
documentation added by this commit.

For #1658.