types: refactor `Type.String` for texts, arrays and maps
This change removes the character and byte size from text types — so
instead of `"text(300 chars)"`, it now returns simply `"text"`.

It also updates the `Type.String` method to include element types for
arrays and maps. Instead of returning just `"array"` or `"map"`, it
now returns `"array(text)"` or `"map(int)"` depending on the element
types.