all: implement data warehouse queries in the MCP server
This commit implements a tool for querying the data warehouse that is
exposed throught the MCP (Model Context Protocol) server.

Note that various aspects of implementation needs to be investigated
more; there are several issues on GitHub that are related to the MCP
server in general and to the tool introduced by this commit. See them
at:

https://github.com/open2b/meergo/issues?q=is%3Aissue%20state%3Aopen%20mcp

Resolves #1658.