assets: discard the JSON path when checking for property existence
This commit ensures that in the action filters, when checking for the
existence of a property inserted inside the filter, the check is not
done by appending any JSON path to the property, since the JSON property
paths are not included in the schema and checking for their existence
will always result in an error.

Close #1664.