assets: prevent opening the combobox list after programmatic focus
This commit restores a combobox behavior that was mistakenly lost with
the commit 7c5c696bd4b9383d4e80e95f9c80f54cd829b820.

Specifically, following the click on one of the options in the combobox
list the combobox input is programmatically put in focus and the cursor
returned to the inside of the expression but the list is not immediately
reopened.

This is necessary because in these cases the user's intent is at most to
continue typing and not to immediately select a new property.