all: extends tools and prompts exposed by the MCP server
This commit makes the MCP server expose more useful information to
satisfy the use case when the user wants to query the data warehouse
data (as events and users) and better understand that data.

So, this commit exposes detailed information about the user and event
schema, their corresponding columns on the data warehouse, exposes some
information about Identity Resolution and user identities, and also adds
a pair of prompts.

Some of these parts will be extended in future commits, as this is just
a step for experimenting and prototyping with the Model Context Protocol
features and their capabilities.

Resolves #1668.