meergo: make `BodyBuffer.Close` non-blocking
Previously, the `Close` method of `BodyBuffer` would block until all
open bodies were closed, in order to return the `*BodyBuffer` to the
pool. However, this behavior prevented proper handling of context
cancellations.

With this change, `Close` no longer waits: the responsibility for
returning the `*BodyBuffer` to the pool is now delegated to the
`Close` method of `bodyReader`—unless all bodies have already been
closed when `Close` is called.