core/filters: fix `Apply` function on `json` properties
This commit addresses two issues affecting the `Apply` function when
used with `json` properties:

- `Apply` incorrectly assumed that condition values for `json`
properties were of type `*state.JSONConditionValue`, while they were
actually of type `state.JSONConditionValue`. This mismatch caused a
panic when applying a filter.

- The `readPropertyFrom` function returned `json.Value(nil)` instead
of `nil` when a path did not exist in the `json` property. As a
result, calling the `Kind` method on `json.Value(nil)` triggered a
panic.

Both issues are now properly handled to prevent runtime panics.