core/events/collector: sync sender with destination settings
Previously, when the settings of a destination app connection on
events were updated, the `waitTime` and `sendEvents` functions used by
the sender were not refreshed. As a result, the sender continued to
use outdated settings.

This change ensures that when the settings are updated, the sender's
`waitTime` and `sendEvents` functions are also properly updated.