core/state: standardize handling of `json.RawMessage` in notifications
This commit standardizes how 'json.RawMessage' fields are handled in
notifications; specifically, it ensures that when a value is missing,
the notification still includes JSON null. This is also useful when
debugging Meergo and inspecting the serialized notifications in the
database.

It's then the keep.go code, after de-serialization, that converts
'json.RawMessage("null")' into 'json.RawMessage(nil)'.