all: change values returned by RawQuery and used by MCP server
This commit causes the warehouse method 'RawQuery' to return a row
iterator. This allows for finer control over memory and thus prevents a
warehouse query from loading an excessive amount of data into memory.

Accordingly, modify the workspace and MCP server to support
sending truncated JSON data to the MCP client.

For #1666.