core/connectors/httpclient: fix `ServeMux.Handler` pattern handling
The HTTP client was relying on the `http.ServeMux.Handler` method to
always return the exact request pattern. However, this assumption was
incorrect.

`Handler` first tries to match the request pattern exactly. If it
doesn’t find one, it applies its own rules to check whether a redirect
target exists. In that case, it returns a redirect handler and only the
path of the target pattern, rather than the original request pattern.

As a result, if a connector made a request with no matching pattern, the
client could end up with either a different pattern or a non-existent
one. In the latter case, the code would panic because `endpointGroup`
would be left as the zero value of its type.

Similarly, when a connector request used an empty path (which should be
treated as "/"), `Handler` could return "/" as the pattern, which may or
may not correspond to a real one.

This change fixes the issue by ensuring that if no exact match exists
for the request (ignoring any redirects), an error is returned instead
of an invalid pattern and a zero endpoint group.

For #1720