core: fix transaction consistency and deadlock issues
Previously, in some cases transactions were started but the queries were
executed on a different connection instead of within the transaction.
As a result, the queries were not atomic or consistent with the
transaction, defeating its purpose. This could also lead to deadlocks
when the maximum number of connections was limited (4 or fewer).

For #1673