connectors/excel: return consistent error for unreadable Excel/ODF files
Previously, `Read` and `Sheets` methods returned an `io.EOF` error when
reading ODF files. Since ODF files are zipped, the error wasn't being
rewritten as it was for non-zipped files.

This change makes the method return a consistent error message
regardless of the underlying issue. Even I/O errors are now wrapped
into this unified error.

It will be up to `core/connectors` to detect I/O errors and rethrow them
(not yet implemented).

So instead of returning `"not a valid Excel '.xlsx' file"` or `"EOF"`,
it now always returns: `"document is not a valid Excel (.xlsx) file or
may be corrupted"`.

Fixes #1746