test: fix tests adding default filter when importing users from events
PR #1712 moved the filter from the collector to the actions, but didn't
update the server-side tests.

So one of the tests (events_test.go) stopped working, because it now
imports identities from any event, which causes the email, which is
required for the test to work, to be overwritten and deleted at some
point by other events which were not imported before.

So this commit adds the default filter where needed, and it also does so
for the tests that coincidentally continued to work, since it's correct
that the behavior hasn't been changed by #1712.