cmd: move event type from path to query string in API endpoints
Previously, the event type was passed to the API endpoint as a path
component. Since an event type can contain any UTF-8 character except
`\x00`, it may also include the sequence `..`.

The `encodeURIComponent` function in JavaScript does not escape `..`,
and even if we manually escape it before calling `fetch`, the browser
still interprets it as `..` and rewrites the URL.

This means there is no safe way to include an arbitrary Unicode string
in a URL path without risking that some characters will be interpreted
in a special way by browsers or servers.

To address this, this commit moves event types from the path to the
query string in API endpoints.