test/meergotester: fix tests hanging forever on Windows
On Windows, it may happen that 'startTime' is exactly equal to 'ts'
because the precision of timestamps is lower: for this reason, it is
necessary to check that 'startTime ≥ ts', not just that it is after.

If this is not done, the check may always be false and the loop
continues indefinitely, blocking the test.

Co-authored-by: Andrea Retini <retini@open2b.com>