all: move file path and table name to query string in API endpoints
Similar to commit 42ed5bb, which moved event types from the path to the query
string of the API endpoints, this commit applies the same approach to file paths
and table names.

These values may contain the `..` sequence that the browsers interpret in a
special way when used in a URL path, even if escaped. Using the query string
ensures that the values can be safely transmitted without being misinterpreted.

Close #1740.