assets: prevent optional properties from being set as identity columns
This commit filters the "readOptional" properties from the list of properties
that can be selected as identity columns.

Furthermore, in the case of JSON files, if a property previously marked as
required is then set to optional, and the new file structure is confirmed, the
identity column of the action is emptied and an error is displayed under the
corresponding input field, thus avoiding a Bad Request error returned by the
server during saving.

Close #1744.