assets, cmd, core: hide password reset if it is not enabled
This commit hides the "Forgot your password?" link from the login page, when the
system is unable to send the password reset email, for instance, when SMTP is
not configured.

Additionally, the Admin now automatically redirects to the login page if the
user navigates to the password reset page but the password reset email is not
enabled.

Close #1739.