all: configure the Filesystem connector via env var instead of setting
This commit changes the Filesystem connector so that the root from which
it reads/writes files is not determined via a setting, but is instead
configured via environment variables.

It also adds an additional option to the connector (this is configured
via env var too) to display a different path in the admin.

These two changes make the connector significantly more usable when used
via Docker, as the user no longer needs to manually configure it;
instead, it is automatically configured by Docker Compose.

This has resulted in several changes to Meergo, including a modification
to the tests that use storage, both server-side and admin-side, and a
change to the way Meergo accesses environment variables (now via
'meergo.GetEnvVars').

Subsequent commits may refine minor details and improve the
documentation, tests, and interface.