core/datastore: skip identity flush if warehouse mode is not normal
This change skips flushing event identities to the data warehouse when
warehouse mode is not set to normal. Previously, identities were
discarded individually when the `EventIdentityWriter.Write` method was
called. However, this meant that a flush could still write to the
warehouse if the mode had changed in the meantime.

For #1349