assets: set correct target when redirecting to connection metrics page
This commit ensures that when the user is redirected to the connection metrics
page, the target is set correctly based on the redirect context.

For instance, if the user clicks on the "Go to metrics for details" link after
the completion of an action targeting users, the metrics page will show user
metrics instead of event metrics.

Closes #1731.