connectors/httpfiles: split the "HTTP Files" connector in two
This commit splits the "HTTP Files" connector in two, so there's an
"HTTP GET" connector that imports and an "HTTP POST" connector that
exports. This makes it much clearer what the connectors do and how to
use them.

Subsequent commits may rename the connectors directory; this hasn't been
done in this commit to minimize the git diff.