core/connectors/httpclient: add request capture via context
This change adds support to the HTTP client for capturing a request and
storing a copy of it in a key provided through the request context.

This feature is particularly useful when testing connectors: a connector
method can be invoked with a value of type `*http.Request` under the
`CaptureRequestContextKey` in the context, and upon return, that value
will be populated with the last request processed by the HTTP client.