core,connectors,meergo: remove `Ack` from `RecordReader` and fix metrics
The `Ack` method from the `meergo.RecordReader` interface was used by
file connectors to signal that a record would not be written to the
file. However, for file exports it is expected that either all records
are written or no file is produced at all. File connectors must
guarantee that all records are written to the file regardless of their
values.

This change removes the `Ack` method and ensures that export metrics are
recorded correctly.

Fixes #1810