core/internal/state: avoid redundant state change notifications
State change notifications must only be sent when the state actually
changes.

However, when linking two connections, in rare cases a notification may
still be triggered even if the connections are already linked. The
`State.linkConnection` method handles this correctly, but the
notification itself is still undesirable. This change prevents the
notification from being sent in such cases.

It also updates `State.unlinkConnection` to skip checking whether the
state has changed. For unlinking, the notification is only sent if the
state really changed.