core: fix panic from writer lifecycle mismatch in collector/datastore
The `Close` method of `datastore.EventIdentityWriter` did not remove the
writer from the store's event identity writers.

As a result, when enabling an action (and thus creating a new event
identity writer), the datastore's `newEventIdentityWriter` would return
an error. This error was not caught by the collector's
`newIdentityWriter`, leaving the collector with a `nil` datastore
writer.

This led to a panic whenever the collector accessed the `nil` writer,
which could happen in several cases:

* the action was disabled
* the action was deleted
* the action's connection was deleted
* event identities were flushed

This change fixes the issue by ensuring closed writers are removed from
the store's event identity writers. However, that alone is not
sufficient, since the initialization and removal of writers in the
collector and the store did not occur within the same state freeze. This
meant the problem could still occur, though less frequently.

To fully resolve the issue, this change makes sure that writer creation
and removal in both the collector and the datastore happen within the
same state freeze.

Fixes #1822