core: return empty slice instead of NotFound in `Workspace.Identities`
This change updates `Workspace.Identities` to return an empty slice when
the user does not exist, instead of a NotFound error.

The method does not distinguish between a non-existent user and a user
with no identities (which can happen if an action is deleted along with
all its identities). Returning an empty slice avoids showing a NotFound
error in the Admin console when clicking the "identities" tab on a user
profile. This behavior is consistent with the "traits" and events tabs,
which do not return errors if the user does not exist.