core/internal/connectors: fix panic on records with invalid columns
This change prevents a panic when reading database records with issues
(invalid `column.Type` with a non-empty `column.Issue`). The panic
occurred because the iterator attempted to access a property in the
action schema using the column name, but such columns may not have a
name.

Additionally, this change zeros columns that have issues or are unused
in the action schema before passing them to the `newDatabaseRecords`
iterator constructor. It also validates the column name returned by the
database connector.

Co-authored-by: Gianluca Mondini <mondini@open2b.com>