core/internal/connectors: skip purge for errored records without ID
Previously, when re-importing, if a record had an error and its ID was
empty, the final purge step would remove it even though it should not.

This change ensures that if a record has an error and its ID is unknown,
the purge is skipped. An error log entry is written in this case.

Additionally, for database connectors, this change prevents returning
the previous record's ID for a record with an error.