doc: refactor SDK documentation by merging SDK doc into connectors
Currently, we have two separate pages for SDK connectors (e.g. Go,
Python, JavaScript, etc.) and the documentation of the SDK itself.

This commit merges that documentation into one, under the 'Connectors'
tab. This results in simpler and intuitive documentation for an user who
wants to send events with their SDK.

Future commits may apply changes to the Developers tab (that now
contains only the connectors development documentation) and make changes
to the pages of the SDK, extending and refining them.