assets: handle `null` values when checking sample fields to highlight
This commit fixes an issue that occurred when viewing samples in full mode
actions, in the case of function-based transformations.

Previously, when calculating which sample fields to highlight based on the
user's selected input paths, a `null` value was mistakenly treated as an object.
This caused the algorithm to look for subfields and highlight fields that the
user had not actually selected.

Closes #1868, #1869