test: make test `TestJavaScriptSDKEventURL` independent from test config
Currently, this test passes only if the tests configuration is the
default; if the user provided another configuration for the tests, as
documented, this test fails.

This is an inconsistency between the tests documentation/configuration
and the actual tests performed.

So, this commit fixes the inconsistent test by making it independent
from the tests configuration.