all: use property paths instead of names for matching properties
Currently, both the input and the output matching properties are limited
to property names, not property paths. This has been acceptable (even if
not ideal) until now. Now, the PR #1865, which refactors the properties
of HubSpot, grouping them within objects, necessarily require the
selection of an output matching property that is a path, not just a
name. So, we decided to make this change.

This commit changes both the input and the output matching property from
property names to property paths.

Resolves #1871.

Co-authored-by: Andrea Retini <retini@open2b.com>
Co-authored-by: Marco Gazerro <gazerro@open2b.com>