assets: make Prettier ignore minified file `snippets/javascript.ts`
Currently, Prettier changes the file ending of the file
'snippets/javascript.ts', in a way that is different from the way that
'npm run minify-snippet' formats it.

So, if these two commands are run in this order:

    npm run prettier
    npm run minify-snippet

There is no problem, but if Prettier is run without running the 'npm run
minify-snippet' after it command, or if their execution order is
inverted, the file is changed respect to the one versioned in the
repository.

So, this commit makes Prettier ignore the 'snippets/javascript.ts',
which should be formatted only by 'npm run minify-snippet'.