meergo: give precedence to env vars over the ones read from '.env'
This commit calls the function 'Load' instead of 'Overload' because we
want the environment variables passed to the process to take precedence
over those read from the '.env' file.

This behavior is already documented in docs.

The choice is mainly driven by the fact that this behavior is the most
expected by a user, since it is the default of Node, Ruby, and Python
libraries that read '.env' files.

Closes #1900.